+++ app/routes/components/FileInput/demos/Main/index.tsx
@@ -1,5 +1,5 @@
 import { FileInput } from "@/components/FileInput";
 
 export default function Demo() {
-  return <FileInput placeholder="Choose file..." />;
+  return <FileInput w="12rem" placeholder="Choose file..." />;
 }