+++ src/components/FileInput/index.ts
@@ -1,4 +1,4 @@
-// FileInput from copy-ui @ 2025-11-21.
+// FileInput from copy-ui @ 2025-12-08.
 
 export type { FileInputProps } from "./file-input";
 export { FileInput } from "./file-input";