+++ src/components/FileInput/CHANGELOG.md
@@ -1,1 +1,2 @@
 - 2025-11-21: Initial version.
+- 2025-12-08: Use `IbsBase` rather than the deprecated `InputBase` component.